% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz3d_utils.R
\name{make_scene}
\alias{make_scene}
\title{Default Plotly 3-D scene layout parameters}
\usage{
make_scene()
}
\value{
A named list for use in \code{plotly::layout(scene = make_scene())}.
}
\description{
Returns a list suitable for the \code{scene} element of \code{plotly::layout()},
with sensible camera angle, axis titles, and equal aspect ratio.
}
\examples{
\dontrun{
library(plotly)
fig <- plot_ly() \%>\%
  plotly::layout(scene = make_scene())
fig
}

}
\seealso{
\code{\link[=panel_layout]{panel_layout()}}
}
