% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_SPI}
\alias{compute_SPI}
\title{SPI (Clarkson method)}
\usage{
compute_SPI(dx, dy, dz)
}
\arguments{
\item{dx}{Numeric vector. X-components of the direction vectors.}

\item{dy}{Numeric vector. Y-components of the direction vectors.}

\item{dz}{Numeric vector. Z-components of the direction vectors.}
}
\value{
A single numeric value in \eqn{[0, 1]}.  Values close to 1
indicate strong preferred orientation; values close to 0 indicate
isotropy.
}
\description{
Computes the SPI \eqn{R} for a set of 3-D scar direction
vectors following the Clarkson convention: the magnitude of the resultant
vector divided by the sum of all individual vector magnitudes.
}
\examples{
# Perfectly aligned vectors along X
compute_SPI(c(1, 1, 1), c(0, 0, 0), c(0, 0, 0))  # returns 1

}
